Fix calculate points function
- create calculate module
- test function from created module
- fix typos
